requirements.txt
aiofiles==0.8.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
autopep8==2.0.2
beautifulsoup4==4.11.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
firebase-admin==5.2.0
google-api-core==2.8.1
google-api-python-client==2.50.0
google-auth==2.7.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.5.2
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
httplib2==0.20.4
idna==3.3
msgpack==1.0.4
multidict==6.0.2
proto-plus==1.20.5
protobuf==3.20.1
py-cord==2.4.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyNaCl==1.4.0
pyparsing==3.0.9
python-dotenv==0.19.2
requests==2.27.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
typing_extensions==4.1.0
uritemplate==4.1.1
urllib3==1.26.9
yarl==1.7.2
